requirements.txt
autobahn==23.1.2
bcrypt==4.0.1
binance-connector==2.0.0
flasgger==0.9.5
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-Track-Usage==2.0.0
gunicorn==20.1.0
marshmallow==3.19.0
marshmallow_enum==1.5.1
marshmallow-sqlalchemy==0.28.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.4
